requirements.txt
pynetbox
zabbix-utils==2.0.1
pyyaml